requirements.txt
sarvamai
python-dotenv
rich
Pillow
